Wheest/gspc dev

+++ python/tvm/auto_scheduler/utils.py
@@ -23,12 +23,15 @@ import multiprocessing.pool
 import queue
 import signal
 import threading
+import traceback
 import os
+
+import numpy as np
 
 try:
     import psutil
 except ImportError:
-    raise ImportError("psutil not found, try `pip install psutil` to fix this")
+    psutil = None
 
 from tvm import rpc
 from tvm.tir import expr
@@ -50,7 +53,7 @@ def get_func_name(func):
     name: str
         The function name.
     """
-    return func.func_name if hasattr(func, 'func_name') else func.__qualname__
+    return func.func_name if hasattr(func, "func_name") else func.__qualname__
 
 
 def get_const_int(exp):
@@ -92,7 +95,6 @@ def get_const_tuple(in_tuple):
     return tuple(get_const_int(x) for x in in_tuple)
 
 
-
 def list_to_tuple(x):
     """ Convert a list to a tuple recursively. """
     assert isinstance(x, list)
@@ -107,7 +109,7 @@ def serialize_args(args):
     ret = []
     for t in args:
         if isinstance(t, Tensor):
-            t = ('TENSOR', get_const_tuple(t.shape), t.dtype)
+            t = ("TENSOR", get_const_tuple(t.shape), t.dtype)
         elif isinstance(t, list):
             t = list_to_tuple(t)
 
@@ -121,68 +123,65 @@ def deserialize_args(args):
     """The inverse function of :code:`serialize_args`"""
     ret = []
     for t in args:
-        if isinstance(t, (tuple, list)) and t[0] == 'TENSOR':
+        if isinstance(t, (tuple, list)) and t[0] == "TENSOR":
             ret.append(placeholder(shape=t[1], dtype=t[2]))
         else:
             ret.append(t)
     return ret
 
 
-class NoDaemonProcess(multiprocessing.Process):
-    @property
-    def daemon(self):
-        return False
-
-    @daemon.setter
-    def daemon(self, value):
-        pass
-
-
-class NoDaemonContext(type(multiprocessing.get_context())):
-    Process = NoDaemonProcess
-
-
-class NoDaemonPool(multiprocessing.pool.Pool):
-    """A no daemon pool version of multiprocessing.Pool.
-    This allows us to start new processings inside the worker function"""
-
-    def __init__(self, *args, **kwargs):
-        kwargs['context'] = NoDaemonContext()
-        super().__init__(*args, **kwargs)
-
-    def __reduce__(self):
-        pass
-
-
 def kill_child_processes(parent_pid, sig=signal.SIGTERM):
     """kill all child processes recursively"""
+    if not psutil:
+        raise ImportError("psutil not found, try `pip install psutil` to fix this")
+
     try:
         parent = psutil.Process(parent_pid)
     except psutil.NoSuchProcess:
         return
-    children = parent.children(recursive=True)
-    for process in children:
-        try:
+
+    try:
+        children = parent.children(recursive=True)
+        for process in children:
             process.send_signal(sig)
-        except psutil.NoSuchProcess:
-            return
+    except psutil.NoSuchProcess:
+        return
 
 
-def call_func_with_timeout(timeout, func, args=(), kwargs=None):
-    """Call a function with timeout"""
-    def func_wrapper(que):
+# The maximum length of traceback information
+MAX_TRACEBACK_INFO_LEN = 512
+
+
+def make_traceback_info():
+    """ Get the error message from traceback. """
+    info = str(traceback.format_exc())
+    if len(info) > MAX_TRACEBACK_INFO_LEN:
+        info = (
+            info[: MAX_TRACEBACK_INFO_LEN // 2] + "\n...\n" + info[-MAX_TRACEBACK_INFO_LEN // 2 :]
+        )
+    return info
+
+
+def _func_wrapper(que, func, args, kwargs):
+    """Call function and return the result over the queue."""
+    try:
         if kwargs:
             que.put(func(*args, **kwargs))
         else:
             que.put(func(*args))
+    # pylint: disable=broad-except
+    except Exception:
+        que.put(Exception(make_traceback_info()))
 
+
+def call_func_with_timeout(timeout, func, args=(), kwargs=None):
+    """Call a function with timeout"""
     que = multiprocessing.Queue(2)
-    process = multiprocessing.Process(target=func_wrapper, args=(que,))
+    process = multiprocessing.Process(target=_func_wrapper, args=(que, func, args, kwargs))
     process.start()
-    process.join(timeout)
 
     try:
-        res = que.get(block=False)
+        res = que.get(timeout=timeout)
     except queue.Empty:
         res = TimeoutError()
 
@@ -199,7 +198,7 @@ def call_func_with_timeout(timeout, func, args=(), kwargs=None):
 
 
 def request_remote(device_key, host=None, port=None, priority=1, timeout=60):
-    """ Request a remote session.
+    """Request a remote session.
 
     Parameters
     ----------
@@ -222,12 +221,11 @@ def request_remote(device_key, host=None, port=None, priority=1, timeout=60):
         The connected remote RPCSession.
     """
     # connect to the tracker
-    host = host or os.environ['TVM_TRACKER_HOST']
-    port = port or int(os.environ['TVM_TRACKER_PORT'])
+    host = host or os.environ["TVM_TRACKER_HOST"]
+    port = port or int(os.environ["TVM_TRACKER_PORT"])
 
     tracker = rpc.connect_tracker(host, port)
-    remote = tracker.request(device_key, priority=priority,
-                             session_timeout=timeout)
+    remote = tracker.request(device_key, priority=priority, session_timeout=timeout)
     return remote
 
 
@@ -259,7 +257,54 @@ def check_remote(device_key, host=None, port=None, priority=100, timeout=10):
     def _check():
         request_remote(device_key, host, port, priority)
 
-    t = threading.Thread(target=_check, )
+    t = threading.Thread(
+        target=_check,
+    )
     t.start()
     t.join(timeout)
     return not t.is_alive()
+
+
+def array_mean(arr):
+    """Compute mean of the elments in a TVM Array<PrimExpr>
+
+    Parameters
+    ----------
+    arr: Array
+        A TVM Array<PrimExpr>
+
+    Returns
+    -------
+    mean: float
+        The mean of the elements in the array
+    """
+    return sum(x.value for x in arr) / len(arr)
+
+
+def to_str_round(x, decimal=6):
+    """Convert an object to str and round float numbers
+
+    Parameters
+    ----------
+    x: Union[str, list, int, float, np.ndarray]
+        The input object
+    decimal: int
+        The precision of decimal fraction
+
+    Returns
+    -------
+    ret: str
+        The string format of these objects
+    """
+    if isinstance(x, str):
+        return x
+    if isinstance(x, (list, tuple, np.ndarray)):
+        return "[" + ", ".join([to_str_round(y, decimal=decimal) for y in x]) + "]"
+    if isinstance(x, dict):
+        return str({k: to_str_round(v) for k, v in x.items()})
+    if isinstance(x, int):
+        return str(x)
+    if isinstance(x, (np.float32, np.float64, float)):
+        format_str = "%%.%df" % decimal
+        return format_str % x
+    raise ValueError("Invalid value: " + str(x) + "\ttype: " + str(type(x)))
